chore: add dedicated service modules
